+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Now offers a shorthand syntax to define colorbar labels and colors via a single mapping.
     - Fixes subtle errors in the handling of parameters and the resulting output.
     - Adds a `ColorManager`-specific documentation page, including an integration example.
+    - Allows creating color palettes using seaborn.
 - !314 fixes an error where facet grid plots did not correctly display the passed data in an error message.
 
 ## v0.18.4